Convert to Singletons and Begin Base File
Removed CommandBase in order to switch to a system that has each
subsystem having a GetInstance function that will return the correct
instance.

Created a SideDrive "subsystem" That will be the base for all other
driving subsystems that involve driving sideways, containing a usefull
set of functions to facilitate working with them.