Initial SwerveDrive Setup
Created SwerveDrive, with an inner class that models a single drive module. Defined several of the functions that will be useful for the DriveModule and the SwerveDrive, though many are as of yet unimplemented. The key basic drive was implemented, though the unfinished constructor for DriveModule makes it not usable.

RobotMap was updated with constants for CAN IDs and DIO ports.